+++ lib/gradients.dart
@@ -1,7 +1,7 @@
-/// Gradients
-///
 /// Suitable classes for working with Flutter's Gradient Apis.
 library gradients;
 
 export 'src/gradient.dart';
-export 'src/gradient_immu.dart';
+export 'src/gradient_immut.dart';
+export 'src/gradient_swatch.dart';
+export 'src/material/grad_of.dart';