+++ lib/src/material/black/black.dart
@@ -18,20 +18,20 @@ class Black extends PaletteRGB {
   const Black() : super.black();
 
   /// 87% opacity.
-  const Black.opacity87() : super.black(opacity:0xDD);
+  const Black.opacity87() : super.black(opacity: 0xDD);
 
   /// 54% opacity.
-  const Black.opacity54() : super.black(opacity:0x8A);
+  const Black.opacity54() : super.black(opacity: 0x8A);
 
   /// 45% opacity.
-  const Black.opacity45() : super.black(opacity:0x73);
+  const Black.opacity45() : super.black(opacity: 0x73);
 
   /// 38% opacity.
-  const Black.opacity38() : super.black(opacity:0x61);
+  const Black.opacity38() : super.black(opacity: 0x61);
 
   /// 26% opacity.
-  const Black.opacity26() : super.black(opacity:0x42);
+  const Black.opacity26() : super.black(opacity: 0x42);
 
   /// 12% opacity.
-  const Black.opacity12() : super.black(opacity:0x1F);
+  const Black.opacity12() : super.black(opacity: 0x1F);
 }