+++ CHANGELOG.md
@@ -7,12 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+
+- `GradOf` class with the constructors `blueGrey`, `grey`, and `brown`.
+- `GradientSwatch` helper abstract class.
+
 ### Changed
 
 - stricter lint rules.
-- improvements in README.
-- Rename 'o' parameter to 'opacity' in `white` and `black` constructors of the
-  `PaletteRGB` class — **BREAKING CHANGES**.
+- improvements to README file.
+- rename 'o' parameter to 'opacity' in `white` and `black` constructors of the
+  `PaletteRGB` class — **BREAKING CHANGE**.
+- rename `GradientImmu` class to `GradientImmut` — **BREAKING CHANGE**.
 
 ### Removed